refactor: - Move template-checking code from DateField to TemplatingPluginTools
